tasks/timezone.yml: Fix TZ handling on Debian
- Previous tests were using America/New_York and I'm using Europe/Paris
so everything is fine, but using US/Central will be changed by
tzdata reconfiguration to 'America/Chicago'. While relying on symlinks
for this one will work, it wont work on other cases like
America/Argentina/ComodRivadavia.

The only way I found is to reuse the function from tzdata package
used to convert the TZ. It's fragile but seems to work.

- Create /etc/timezone only if it already exists and call
  dpkg-reconfigure tzdata to configure the timezone, as mentionned
  in the debian bugs #813226, #848143.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>