tasks/timezone.yml: Fix wrong /etc/localtime mode
ansible-lint was issuing an error as owner/group/mode was not set,
so they've been added in "tasks/timezone.yml: Fix ansible-lint messages".
The settings used are a copy of what was checked by the verifier.

This was forgetting that in case of symlink, file will follow it
when owner/group/mode is set and set the target. So the correct
mode has to be 0644 and not 0777.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>