tasks/timezone.yml: Debian: Trim script: module output
It turns out that if the role is configured to be run as root
(for instance, with 'become: yes' on the play), this will result
in adding an extra empty line in the output, leading to timezone
being set to an empty string. Change the code to use the full
stdout and use trim to remove the '\r\n'.

Since the docker test is connecting as root, this issue was not
visible. Reproduce this issue by using vagrant-libvirt.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>